+++ packages/liveblocks-react-ui/src/components/Comment.tsx
@@ -18,6 +18,7 @@ import type {
   FormEvent,
   MouseEvent,
   ReactNode,
+  RefObject,
   SyntheticEvent,
 } from "react";
 import React, {
@@ -56,6 +57,7 @@ import { MENTION_CHARACTER } from "../slate/plugins/mentions";
 import { classNames } from "../utils/class-names";
 import { useRefs } from "../utils/use-refs";
 import { useVisibleCallback } from "../utils/use-visible";
+import { useWindowFocus } from "../utils/use-window-focus";
 import { Composer } from "./Composer";
 import { Avatar } from "./internal/Avatar";
 import { Button } from "./internal/Button";
@@ -127,7 +129,7 @@ export interface CommentProps extends ComponentPropsWithoutRef<"div"> {
   /**
    * @internal
    */
-  markThreadAsReadWhenVisible?: string;
+  autoMarkReadThreadId?: string;
 
   /**
    * @internal
@@ -323,6 +325,34 @@ export const CommentNonInteractiveReaction = forwardRef<
   );
 });
 
+// A void component (which doesn't render anything) responsible for marking a thread
+// as read when the comment it's used in becomes visible.
+// Moving this logic into a separate component allows us to use the visibility
+// and focus hooks "conditionally" by conditionally rendering this component.
+function AutoMarkReadThreadIdHandler({
+  threadId,
+  commentRef,
+}: {
+  threadId: string;
+  commentRef: RefObject<HTMLElement>;
+}) {
+  const markThreadAsRead = useMarkThreadAsRead();
+  const isWindowFocused = useWindowFocus();
+
+  useVisibleCallback(
+    commentRef,
+    () => {
+      markThreadAsRead(threadId);
+    },
+    {
+      // The underlying IntersectionObserver is only enabled when the window is focused
+      enabled: isWindowFocused,
+    }
+  );
+
+  return null;
+}
+
 /**
  * Displays a single comment.
  *
@@ -349,7 +379,7 @@ export const Comment = forwardRef<HTMLDivElement, CommentProps>(
       className,
       additionalActions,
       additionalActionsClassName,
-      markThreadAsReadWhenVisible,
+      autoMarkReadThreadId,
       ...props
     },
     forwardedRef
@@ -361,22 +391,11 @@ export const Comment = forwardRef<HTMLDivElement, CommentProps>(
     const editComment = useEditComment();
     const addReaction = useAddReaction();
     const removeReaction = useRemoveReaction();
-    const markThreadAsRead = useMarkThreadAsRead();
     const $ = useOverrides(overrides);
     const [isEditing, setEditing] = useState(false);
     const [isTarget, setTarget] = useState(false);
     const [isMoreActionOpen, setMoreActionOpen] = useState(false);
     const [isReactionActionOpen, setReactionActionOpen] = useState(false);
-
-    const markVisibleThreadAsRead = useCallback(() => {
-      if (markThreadAsReadWhenVisible) {
-        markThreadAsRead(markThreadAsReadWhenVisible);
-      }
-    }, [markThreadAsRead, markThreadAsReadWhenVisible]);
-
-    useVisibleCallback(ref, markVisibleThreadAsRead, {
-      enabled: Boolean(markThreadAsReadWhenVisible),
-    });
 
     const stopPropagation = useCallback((event: SyntheticEvent) => {
       event.stopPropagation();
@@ -481,6 +500,12 @@ export const Comment = forwardRef<HTMLDivElement, CommentProps>(
 
     return (
       <TooltipProvider>
+        {autoMarkReadThreadId && (
+          <AutoMarkReadThreadIdHandler
+            commentRef={ref}
+            threadId={autoMarkReadThreadId}
+          />
+        )}
         <div
           id={comment.id}
           className={classNames(